+++ CHANGES.md
@@ -8,7 +8,7 @@ Changes in 7.8
 - Remove Reset button from Find a Student / User forms in Students/Search.inc.php & Users/Search.inc.php
 - CSS & JS open submenu on focus & focus-within in warehouse.js, stylesheet.css & rtl.css
 - CSS menu link & button color on focus in stylesheet.css & colors.css
-- Fix check students Course Status in PrintClassLists.php, PrintClassPictures.php, ClassSearchWidget.fnc.php & Referrals.php
+- Fix check students Course Status in PrintClassLists.php, PrintClassPictures.php, ClassSearchWidget.fnc.php, Referrals.php & EmailReferral.fnc.php
 - Add Include Inactive Students checkbox in MasterScheduleReport.php & RequestsReport.php
 - Fix unset current student (check Course Status) when MP updated in Side.php